Refactored script.js to utilise modules, imports and exports for organisational purposes
