+++ public/html/index.html
@@ -57,9 +57,9 @@
 
     <div class="toolbar">
         <ul>
-            <li><button>Tool 1</button></li>
-            <li><button>Tool 2</button></li>
-            <li><button>Tool 3</button></li>
+            <li><button id="cropButton">Crop</button></li>
+            <li><button id="rotateButton">Rotate</button></li>
+            <li><button id="brightnessButton">Brightness</button></li>
             <li><button>Tool 4</button></li>
             <li><button>Tool 5</button></li>
         </ul>
@@ -102,5 +102,5 @@
     </div>
 </div>
 </body>
-<script src="../js/script.js"></script>
+<script type="module" src="../js/script.js"></script>
 </html>
